backend/requirements.txt
# Web Framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0

# Validation & Configuration
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Environment
python-dotenv>=1.0.0,<2.0.0

# Google Gemini SDK
google-genai>=1.0.0

# HTTP client for testing
httpx>=0.26.0
